# openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: info@mod.construction
    name: Contact us
    url: https://mod.construction
  description: "mod-dlm is an open source domain specific language to describe prefab\
    \ construction elements. This project provides tools for defining entity models\
    \ and automatically generating CRUD APIs, enabling suppliers to create and manage\
    \ catalogs of their products easily."
  license:
    name: MIT
    url: https://github.com/mod-construction/mod-dlm/blob/main/LICENSE
  title: mod-dlm API
  version: 1.0.8
  x-logo:
    url: mod-dlm-logo.png
    backgroundColor: '#FFFFFF'
    altText: mod-dlm logo
servers:
- description: The dev server.
  url: http://localhost:8080
- description: The production server.
  url: https://mod.construction
security:
- apiKey: []
  bearer: []
paths:
  /element:
    get:
      description: Gets an element from the database.
      operationId: getElement
      parameters:
      - description: A unique identifier for an element
        explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/ElementId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrefabElement'
          description: The element was retrieved successfully.
        "400":
          description: Bad request.
        "401":
          description: Authorization information is missing or invalid.
        "404":
          description: An element with the specified ID was not found.
      summary: Get an element
    post:
      description: Creates a new element in the database.
      operationId: createElement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementCreate'
        description: The element to create.
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrefabElement'
          description: The element was created successfully.
        "400":
          description: Bad request.
        "401":
          description: Authorization information is missing or invalid.
      summary: Create a new element
components:
  schemas:
    ElementId:
      description: A unique identifier for an element
      example: 4dd643ff-7ec7-4666-9c88-50b7d3da34e4
      format: uuid
      title: id
      type: string
    PrefabElement:
      description: Prefab Element
      example:
        boundingBox:
          depth: 1.4658129805029452
          width: 0.8008281904610115
          height: 6.027456183070403
        images:
        - images
        - images
        installationAndConnectivity:
          installationTime: installationTime
          connectionType: Bolt-on
          compatibility: Steel Frame
        dimensional:
          width:
            min: 5.962133916683182
            max: 5.637376656633329
          length:
            min: 5.962133916683182
            max: 5.637376656633329
          height:
            min: 5.962133916683182
            max: 5.637376656633329
        structuralProperties:
          windLoadResistance: Class A
          loadBearingCapacity:
            unit: "N"
            loadDistribution: loadDistribution
            maximumLoad: 2.3021358869347655
          seismicResistance: Zone 1
        lifecycleAndMaintenance:
          maintenanceRequirements: maintenanceRequirements
          warranty: warranty
          expectedLifespan: expectedLifespan
        description: description
        documentationAndCompliance:
          technicalSpecifications: technicalSpecifications
          certifications:
            certificationTypes:
            - certificationTypes
            - certificationTypes
            regulatoryApprovals:
            - regulatoryApprovals
            - regulatoryApprovals
        sustainability:
          countryOfManufacturing: countryOfManufacturing
          recyclability: 0%
          classification: A+
          energyEfficiency: R-1
          VOCEmissions: None
        productCategory:
          category: Solid Wall Panels
          buildingSystem: Wall
        aestheticAndCustomizationOptions:
          colorOptions:
          - colorOptions
          - colorOptions
          texture: texture
          modularAdaptability: modularAdaptability
        performance:
          thermalTransmittance: 1.4122802483006218
          acousticProperties:
            acousticPerformance: acousticPerformance
            soundInsulationRating: soundInsulationRating
          resistanceToFireClassification: A1
        material:
          finishMaterial: Timber
          structuralMaterial: Timber
        name: name
        id: 4dd643ff-7ec7-4666-9c88-50b7d3da34e4
        economicFactors:
          manufacturingLeadTime: manufacturingLeadTime
          costPerUnit: 9.301444243932576
      properties:
        id:
          description: A unique identifier for an element
          example: 4dd643ff-7ec7-4666-9c88-50b7d3da34e4
          format: uuid
          title: id
          type: string
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        boundingBox:
          $ref: '#/components/schemas/PrefabElement_boundingBox'
        images:
          items:
            type: string
          title: images
          type: array
        productCategory:
          $ref: '#/components/schemas/ProductCategory'
        material:
          $ref: '#/components/schemas/PrefabElement_material'
        dimensional:
          $ref: '#/components/schemas/PrefabElement_dimensional'
        structuralProperties:
          $ref: '#/components/schemas/PrefabElement_structuralProperties'
        performance:
          $ref: '#/components/schemas/PrefabElement_performance'
        sustainability:
          $ref: '#/components/schemas/PrefabElement_sustainability'
        documentationAndCompliance:
          $ref: '#/components/schemas/PrefabElement_documentationAndCompliance'
        installationAndConnectivity:
          $ref: '#/components/schemas/PrefabElement_installationAndConnectivity'
        lifecycleAndMaintenance:
          $ref: '#/components/schemas/PrefabElement_lifecycleAndMaintenance'
        aestheticAndCustomizationOptions:
          $ref: '#/components/schemas/PrefabElement_aestheticAndCustomizationOptions'
        economicFactors:
          $ref: '#/components/schemas/PrefabElement_economicFactors'
      required:
      - boundingBox
      - description
      - dimensional
      - id
      - images
      - material
      - name
      - productCategory
      title: PrefabElement
    ProductCategory:
      description: Product Category
      example:
        category: Solid Wall Panels
        buildingSystem: Wall
      properties:
        category:
          enum:
          - Boarding
          - Solid Wall Panels
          - Closed Wall Panels
          - Twinwall
          - Open Wall Panels
          - Structural Insulated Panels (SIPs)
          - Insulated Concrete Panels
          - Prefabricated Balcony
          - Pod
          - Whole Building System
          - Structural Frame
          - Facade System
          - Hollowcore Floor
          - Concrete Lattice Floor
          - Floor Cassettes
          - Solid Floor Panels
          - Volumetric module
          - Prefabricated Plant
          - Roof Panel
          - Roof Truss
          - Prefabricated Stairs
          type: string
        buildingSystem:
          enum:
          - Wall
          - Balcony
          - Pod
          - Frame
          - Facade
          - Floors
          - Modules
          - Plants
          - Roofs
          - Stairs
          type: string
      required:
      - buildingSystem
      - category
      title: ProductCategory
    ElementCreate:
      description: An element to create.
      example:
        boundingBox:
          depth: 1.4658129805029452
          width: 0.8008281904610115
          height: 6.027456183070403
        images:
        - images
        - images
        installationAndConnectivity:
          installationTime: installationTime
          connectionType: Bolt-on
          compatibility: Steel Frame
        dimensional:
          width:
            min: 5.962133916683182
            max: 5.637376656633329
          length:
            min: 5.962133916683182
            max: 5.637376656633329
          height:
            min: 5.962133916683182
            max: 5.637376656633329
        structuralProperties:
          windLoadResistance: Class A
          loadBearingCapacity:
            unit: "N"
            loadDistribution: loadDistribution
            maximumLoad: 2.3021358869347655
          seismicResistance: Zone 1
        lifecycleAndMaintenance:
          maintenanceRequirements: maintenanceRequirements
          warranty: warranty
          expectedLifespan: expectedLifespan
        description: description
        documentationAndCompliance:
          technicalSpecifications: technicalSpecifications
          certifications:
            certificationTypes:
            - certificationTypes
            - certificationTypes
            regulatoryApprovals:
            - regulatoryApprovals
            - regulatoryApprovals
        sustainability:
          countryOfManufacturing: countryOfManufacturing
          recyclability: 0%
          classification: A+
          energyEfficiency: R-1
          VOCEmissions: None
        productCategory:
          category: Solid Wall Panels
          buildingSystem: Wall
        aestheticAndCustomizationOptions:
          colorOptions:
          - colorOptions
          - colorOptions
          texture: texture
          modularAdaptability: modularAdaptability
        performance:
          thermalTransmittance: 1.4122802483006218
          acousticProperties:
            acousticPerformance: acousticPerformance
            soundInsulationRating: soundInsulationRating
          resistanceToFireClassification: A1
        material:
          finishMaterial: Timber
          structuralMaterial: Timber
        name: name
        economicFactors:
          manufacturingLeadTime: manufacturingLeadTime
          costPerUnit: 9.301444243932576
      properties:
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        boundingBox:
          $ref: '#/components/schemas/PrefabElement_boundingBox'
        images:
          items:
            type: string
          title: images
          type: array
        productCategory:
          $ref: '#/components/schemas/ProductCategory'
        material:
          $ref: '#/components/schemas/PrefabElement_material'
        dimensional:
          $ref: '#/components/schemas/PrefabElement_dimensional'
        structuralProperties:
          $ref: '#/components/schemas/PrefabElement_structuralProperties'
        performance:
          $ref: '#/components/schemas/PrefabElement_performance'
        sustainability:
          $ref: '#/components/schemas/PrefabElement_sustainability'
        documentationAndCompliance:
          $ref: '#/components/schemas/PrefabElement_documentationAndCompliance'
        installationAndConnectivity:
          $ref: '#/components/schemas/PrefabElement_installationAndConnectivity'
        lifecycleAndMaintenance:
          $ref: '#/components/schemas/PrefabElement_lifecycleAndMaintenance'
        aestheticAndCustomizationOptions:
          $ref: '#/components/schemas/PrefabElement_aestheticAndCustomizationOptions'
        economicFactors:
          $ref: '#/components/schemas/PrefabElement_economicFactors'
      required:
      - boundingBox
      - description
      - dimensional
      - images
      - material
      - name
      - productCategory
      title: ElementCreate
    PrefabElement_boundingBox:
      description: The bounding box of an element in millimeters.
      example:
        depth: 1.4658129805029452
        width: 0.8008281904610115
        height: 6.027456183070403
      properties:
        width:
          title: width
          type: number
        height:
          title: height
          type: number
        depth:
          title: depth
          type: number
      required:
      - depth
      - height
      - width
      title: PrefabElement_boundingBox
    PrefabElement_material:
      example:
        finishMaterial: Timber
        structuralMaterial: Timber
      properties:
        finishMaterial:
          enum:
          - Timber
          - Steel
          - Concrete
          - Hybrid
          - Other
          title: finishMaterial
          type: string
        structuralMaterial:
          enum:
          - Timber
          - Steel
          - Concrete
          - Hybrid
          - Other
          title: structuralMaterial
          type: string
      required:
      - finishMaterial
      - structuralMaterial
      title: PrefabElement_material
    PrefabElement_dimensional_width:
      example:
        min: 5.962133916683182
        max: 5.637376656633329
      properties:
        min:
          title: min
          type: number
        max:
          title: max
          type: number
      required:
      - max
      - min
      title: PrefabElement_dimensional_width
    PrefabElement_dimensional:
      example:
        width:
          min: 5.962133916683182
          max: 5.637376656633329
        length:
          min: 5.962133916683182
          max: 5.637376656633329
        height:
          min: 5.962133916683182
          max: 5.637376656633329
      properties:
        width:
          $ref: '#/components/schemas/PrefabElement_dimensional_width'
        height:
          $ref: '#/components/schemas/PrefabElement_dimensional_width'
        length:
          $ref: '#/components/schemas/PrefabElement_dimensional_width'
      required:
      - height
      - length
      - width
      title: PrefabElement_dimensional
    PrefabElement_structuralProperties_loadBearingCapacity:
      example:
        unit: "N"
        loadDistribution: loadDistribution
        maximumLoad: 2.3021358869347655
      properties:
        maximumLoad:
          title: maximumLoad
          type: number
        unit:
          enum:
          - "N"
          - kN
          - lbf
          - kgf
          title: unit
          type: string
        loadDistribution:
          title: loadDistribution
          type: string
      required:
      - loadDistribution
      - maximumLoad
      - unit
      title: PrefabElement_structuralProperties_loadBearingCapacity
    PrefabElement_structuralProperties:
      example:
        windLoadResistance: Class A
        loadBearingCapacity:
          unit: "N"
          loadDistribution: loadDistribution
          maximumLoad: 2.3021358869347655
        seismicResistance: Zone 1
      properties:
        loadBearingCapacity:
          $ref: '#/components/schemas/PrefabElement_structuralProperties_loadBearingCapacity'
        seismicResistance:
          enum:
          - Zone 1
          - Zone 2
          - Zone 3
          - Zone 4
          title: seismicResistance
          type: string
        windLoadResistance:
          enum:
          - Class A
          - Class B
          - Class C
          - Class D
          title: windLoadResistance
          type: string
      title: PrefabElement_structuralProperties
    PrefabElement_performance_acousticProperties:
      example:
        acousticPerformance: acousticPerformance
        soundInsulationRating: soundInsulationRating
      properties:
        soundInsulationRating:
          title: soundInsulationRating
          type: string
        acousticPerformance:
          title: acousticPerformance
          type: string
      required:
      - acousticPerformance
      - soundInsulationRating
      title: PrefabElement_performance_acousticProperties
    PrefabElement_performance:
      example:
        thermalTransmittance: 1.4122802483006218
        acousticProperties:
          acousticPerformance: acousticPerformance
          soundInsulationRating: soundInsulationRating
        resistanceToFireClassification: A1
      properties:
        resistanceToFireClassification:
          enum:
          - A1
          - A2
          - B
          - C
          - D
          - E
          - F
          - 1-hour
          - 2-hour
          - 3-hour
          - 4-hour
          title: resistanceToFireClassification
          type: string
        thermalTransmittance:
          maximum: 2
          minimum: 0
          title: thermalTransmittance
          type: number
        acousticProperties:
          $ref: '#/components/schemas/PrefabElement_performance_acousticProperties'
      required:
      - resistanceToFireClassification
      - thermalTransmittance
      title: PrefabElement_performance
    PrefabElement_sustainability:
      example:
        countryOfManufacturing: countryOfManufacturing
        recyclability: 0%
        classification: A+
        energyEfficiency: R-1
        VOCEmissions: None
      properties:
        countryOfManufacturing:
          title: countryOfManufacturing
          type: string
        classification:
          enum:
          - A+
          - A
          - B
          - C
          - D
          title: classification
          type: string
        VOCEmissions:
          enum:
          - None
          - Low
          - Medium
          - High
          title: VOCEmissions
          type: string
        recyclability:
          enum:
          - 0%
          - 25%
          - 50%
          - 75%
          - 100%
          title: recyclability
          type: string
        energyEfficiency:
          enum:
          - R-1
          - R-2
          - R-5
          - Energy Star
          - LEED Certified
          - LEED Silver
          - LEED Gold
          - LEED Platinum
          title: energyEfficiency
          type: string
      required:
      - classification
      - countryOfManufacturing
      title: PrefabElement_sustainability
    PrefabElement_documentationAndCompliance_certifications:
      example:
        certificationTypes:
        - certificationTypes
        - certificationTypes
        regulatoryApprovals:
        - regulatoryApprovals
        - regulatoryApprovals
      properties:
        certificationTypes:
          items:
            type: string
          title: certificationTypes
          type: array
        regulatoryApprovals:
          items:
            type: string
          title: regulatoryApprovals
          type: array
      required:
      - certificationTypes
      - regulatoryApprovals
      title: PrefabElement_documentationAndCompliance_certifications
    PrefabElement_documentationAndCompliance:
      example:
        technicalSpecifications: technicalSpecifications
        certifications:
          certificationTypes:
          - certificationTypes
          - certificationTypes
          regulatoryApprovals:
          - regulatoryApprovals
          - regulatoryApprovals
      properties:
        technicalSpecifications:
          title: technicalSpecifications
          type: string
        certifications:
          $ref: '#/components/schemas/PrefabElement_documentationAndCompliance_certifications'
      required:
      - certifications
      - technicalSpecifications
      title: PrefabElement_documentationAndCompliance
    PrefabElement_installationAndConnectivity:
      example:
        installationTime: installationTime
        connectionType: Bolt-on
        compatibility: Steel Frame
      properties:
        connectionType:
          enum:
          - Bolt-on
          - Welded
          - Clip-on
          - Adhesive
          - Mortar
          title: connectionType
          type: string
        installationTime:
          title: installationTime
          type: string
        compatibility:
          enum:
          - Steel Frame
          - Wood Frame
          - Concrete Structure
          - Brickwork
          - Modular Systems
          - Glass Facades
          - Composite Materials
          title: compatibility
          type: string
      title: PrefabElement_installationAndConnectivity
    PrefabElement_lifecycleAndMaintenance:
      example:
        maintenanceRequirements: maintenanceRequirements
        warranty: warranty
        expectedLifespan: expectedLifespan
      properties:
        expectedLifespan:
          title: expectedLifespan
          type: string
        maintenanceRequirements:
          title: maintenanceRequirements
          type: string
        warranty:
          title: warranty
          type: string
      title: PrefabElement_lifecycleAndMaintenance
    PrefabElement_aestheticAndCustomizationOptions:
      example:
        colorOptions:
        - colorOptions
        - colorOptions
        texture: texture
        modularAdaptability: modularAdaptability
      properties:
        colorOptions:
          items:
            type: string
          title: colorOptions
          type: array
        texture:
          title: texture
          type: string
        modularAdaptability:
          title: modularAdaptability
          type: string
      title: PrefabElement_aestheticAndCustomizationOptions
    PrefabElement_economicFactors:
      example:
        manufacturingLeadTime: manufacturingLeadTime
        costPerUnit: 9.301444243932576
      properties:
        costPerUnit:
          title: costPerUnit
          type: number
        manufacturingLeadTime:
          title: manufacturingLeadTime
          type: string
      title: PrefabElement_economicFactors
  securitySchemes:
    apiKey:
      description: API key for programmatic access.
      in: header
      name: X-API-KEY
      type: apiKey
    bearer:
      bearerFormat: JWT
      description: An embedded jason web token.
      scheme: bearer
      type: http